+++ box/business_processes/import_bank_statement.rb
@@ -6,7 +6,6 @@ require_relative '../models/account'
 require_relative '../models/bank_statement'
 require_relative '../../lib/checksum_generator'
 
-
 # more general matching regex that covers both newlines and newlines with dashes
 Cmxl.config[:statement_separator] = /(\n-?)(?=:20)/m