+++ box/models/statement.rb
@@ -45,7 +45,6 @@ module Box
     end
 
     class << self
-
       def generic_filter(query, account_id: nil, transaction_id: nil, from: nil, to: nil, type: nil, **_unused)
         # Filter by account id
         query = query.where(account_id: account_id) if account_id.present?
@@ -84,6 +83,10 @@ module Box
 
     def type
       debit? ? 'debit' : 'credit'
+    end
+
+    def amount_in_cents
+      amount
     end
 
     def as_event_payload