+++ migrations/20190529173000_add_sha_to_bank_statement.rb
@@ -4,6 +4,10 @@ Sequel.migration do
   up do
     add_column :bank_statements, :sha, String
     add_index :bank_statements, :sha
+
+    require 'rake'
+    load 'Rakefile'
+    Rake::Task['migration_tasks:calculate_bank_statements_sha'].invoke
   end
 
   down do