+++ migrations/20190705151345_adds_partner_ebics_users.rb
@@ -3,6 +3,10 @@
 Sequel.migration do
   up do
     add_column :ebics_users, :partner, String
+
+    require 'rake'
+    load 'Rakefile'
+    Rake::Task['migration_tasks:copy_partners'].invoke
   end
 
   down do