+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.20
+
+- `[ENHANCEMENT]` Use transaction-id to calculate checksum if available
+- `[HOUSEKEEPING]` Migration will recalculate checksums for statements
+
 ## 1.3.19
 
 - `[ENHANCEMENT]` Adds transactions detail endpoint to v2 api